sitemap: rafactor, moved it to separate files
